+++ src/blocks/event-details/index.tsx
@@ -1,6 +1,7 @@
 import { registerBlockType } from '@wordpress/blocks';
-import { currencyDollar } from '@wordpress/icons';
+import { button } from '@wordpress/icons';
 import './style.scss';
+
 /**
  * Internal dependencies
  */
@@ -8,7 +9,7 @@ import block from './block.json';
 import Edit from './Edit';
 
 registerBlockType( block.name, {
-	icon: currencyDollar,
+	icon: button,
 	edit: Edit,
 	save: () => null,
 } );